+++ web-components/src/components/cards/card.stories.tsx
@@ -14,6 +14,7 @@ import CurrentCreditsIcon from 'web-components/lib/components/icons/CurrentCredi
 import ProjectCard from 'web-components/lib/components/cards/ProjectCard';
 import GreenCard from 'web-components/lib/components/cards/GreenCard';
 import GreenTopIconCard from 'web-components/lib/components/cards/GreenTopIconCard';
+import CreateProjectPlanCard from 'web-components/lib/components/cards/CreateProjectPlanCard';
 
 export default {
   title: 'Components|Cards',
@@ -166,3 +167,11 @@ export const greenTopIconCard = (): JSX.Element => (
 );
 
 export const onBoardingCard = (): JSX.Element => <OnBoardingCard>some content</OnBoardingCard>;
+
+export const createProjectPlanCard = (): JSX.Element => (
+  <CreateProjectPlanCard isFirstProject onClick={() => {}} />
+);
+
+export const createProjectPlanCardAddAnother = (): JSX.Element => (
+  <CreateProjectPlanCard isFirstProject={false} onClick={() => {}} />
+);
